chore: extracted the configurable state to its own file
