+++ src/store/configuration/configurationState.js
@@ -1,0 +1,14 @@
+import { bbmriConfig } from '../../config/configManager'
+
+const config = bbmriConfig()
+
+export const configurationState = {
+  i18n: config.i18n,
+  removeFreemarkerMargin: config.removeFreemarkerMargin,
+  biobankCardShowCollections: config.biobankCardShowCollections,
+  filterMenuInitiallyFolded: config.filterMenuInitiallyFolded,
+  biobankColumns: config.biobankColumns,
+  collectionColumns: config.collectionColumns,
+  filterFacets: config.filterFacets,
+  menuHeight: config.menuHeight // standard menu height in pixels for use in sticky-position
+}
